requirements.txt
# Automated Story Engine Dependencies
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic
python-dotenv==1.0.0

# LLM Integration
openai==1.3.7
anthropic==0.7.8
tiktoken
google-generativeai==0.3.2

# Image Generation
pillow
requests==2.31.0
aiohttp==3.9.1

# Audio/TTS Processing
pydub==0.25.1
# elevenlabs==0.2.26  # Optional - only needed if using ElevenLabs
azure-cognitiveservices-speech==1.34.0  # Optional - only needed if using Azure TTS

# Media Processing
moviepy==1.0.3
ffmpeg-python==0.2.0

# Data & Storage
aiosqlite==0.19.0
redis==5.0.1
numpy==1.26.1
# distutils
pandas>=2.1.4

# Vector Embeddings (for dynamic examples)
sentence-transformers==2.2.2
chromadb==0.4.18

# Utilities
asyncio==3.4.3
jinja2==3.1.2
pyyaml==6.0.1
jsonschema==4.19.2
rich==13.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
